Create resetPeriodicTasks method on custom ExecutorService
During the benchmark, when the number of concurrent workers was increased too
much, it could happen that the workers would execute less frequently than what
they request, and so they could "fall behind" in execution time in such way
to affect the posterior performance of the benchmark (workers would execute
more frequently to compensate for the previous delay). To avoid this, using
newly created Signaler to create a resetPeriodicTasks method to clear the
possibly accumulated delay in the scheduled jobs, and calling it every time
the number of concurrent workers change.