Fix rare StackOverflowError
Error happened on BlockingQueue, when the fetch operation was completed before the
mCurrentOperation field was set, making the callback that clears that field execute
before the field was actually set. That way, the field would be kept set as a completed
future, which executes it's added callbacks synchronously on add, looping indefinitely
on poll() or peek().